Add shared timezone
