+++ tsconfig.base.json
@@ -356,6 +356,9 @@
       "@openops/engine": ["packages/engine/src/main.ts"],
       "@openops/server-shared": ["packages/server/shared/src/index.ts"],
       "@openops/shared": ["packages/shared/src/index.ts"],
+      "@openops/shared/timezones": [
+        "packages/shared/src/lib/common/timezones.ts"
+      ],
       "@openops/ui-kit": ["packages/ui-kit/src/index.ts"],
       "Aminos": ["packages/blocks/Aminos/src/index.ts"],
       "hyperglance": ["packages/blocks/hyperglance/src/index.ts"],
